removed userWeightedText column type
-added userWeightedText function into text colum type
